refactor: output 뷰의 메서드를 기능에 따라 컴포넌트로 분리
